Fix WaitForCompaction data race
The scenario was this:
1. Writer thread A calls WaitForOngoingCompaction
2. Thread B calls db.WaitForCompaction (did not require writer_mutex_)
3. Background compaction finishes, A or B calls get().
4. We now have UB since A or B will call get() on an invalid thread.

We could have alternatively have had each thread call wait(). However,
I don't like the idea of multiple threads wait()-ing on a
single std::future...

A. It's not safe to do that for std::shared_future, so there's no
reason to expect that it's safe for a std::future

B. We have to be VERY sure we NEVER CALL get() while another thread
is wait()-ing!